Updated AppModule DI. Extracted authentication DI to a new module.
